Completely reworked the BaseService. The functions are no longer arrow functions and also they are no longer async, due to a bug in the super call caused by Babel. Instead, they use the old promises and the consumer will be able to handle that as they want.
Fixed an issue where the Snackbar would not appear if it generated from the actions container since that was no longer visible if the user was not hovering over it.

The withSnackbar helper function now also throws the error it gets to avoid swallowing errors.

Implemented the edit functionality for Collections only.