Implemented the reorder functionality for sequences. There are two arrows (up and down) in the table which are visible only when no options table is visible (it works even if they are, but it's confusing).
When clicked, these arrows update the position of the current row and the row found before or after the current one, depending on which button was pressed.

Currently, there are 2 update calls made for each reorder operation, but the backend will eventually support multiple updates only for this special use case.